CityHelplineBackend/image-verification-backend-package/requirements.txt
# Core ML Dependencies
torch>=2.6.0,<2.12
torchvision>=0.21.0,<0.27
numpy>=1.24,<2.2
ultralytics>=8.3,<9.0
PyYAML>=6.0,<7.0

# Image Processing
opencv-python>=4.8,<5.0
Pillow>=10.0,<12.0
Pillow-SIMD==9.3.0.post1; platform_system != 'Windows'

# Utilities
tqdm>=4.65,<5.0
matplotlib>=3.7,<3.11

# Development & Testing
pytest>=7.4,<9.0
pytest-cov>=4.1,<7.0

# Optional: Data handling
pandas>=2.0,<2.4
scikit-learn>=1.3,<1.8
